+++ doc/CHANGELOG.md
@@ -29,8 +29,9 @@ See doc/COPYRIGHT.rdoc for more details.
 
 # Changelog
 
-* `#3337` Fix: Use permitted params in EnumerationsController
 * `#3054` Fix: Some Projects cannot be deleted
+* `#3331` use permitted_params for group_controller
+* `#3337` Fix: Use permitted params in EnumerationsController
 
 ## 3.0.0pre37